Move sketch to an appropriately named folder
The Arduino IDE requires that a sketch be located in a folder of the same name. When opening a file that does not meet this requirement the Arduino IDE presents a dialog:

The file "WiFiSPIESP.ino" needs to be inside a sketch folder named "WiFiSPIESP".
Create this folder, move the file, and continue?

After clicking "OK" the Arduino IDE currently moves only the file WiFiSPIESP.ino to the new folder, leaving behind the other sketch files. This causes compilation of the sketch to fail:

fatal error: SPICalls.h: No such file or directory

The obvious alternative would be to rename the sketch or repository to match on filename case sensitive operating systems such as Linux (Windows is filename case insensitive).  However, it is common for people to use GitHub's Clone or Download > Download ZIP feature to download sketches, rather than doing a Git clone, and GitHub appends the branch/tag/commit to the downloaded folder name, which will cause the folder to still not match the sketch name.